requirements.txt
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
numpy
torch
pyelftools
tqdm
scikit-learn
pandas
